Move InvoiceListPage.
